requirements.txt
streamlit>=1.33.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
langchain>=0.2.16
langchain-core>=0.2.38
langchain-community>=0.2.16
langchain-text-splitters>=0.2.2
langchain-groq>=0.1.9
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.2
chromadb>=0.5.5
sentence-transformers>=3.0.1
pypdf>=4.3.1
python-dotenv>=1.0.1
